refactor(product-review): implement correct folder structure and add missing parts from task
