+++ src/app/[locale]/(website)/products/[id]/_components/comment-component.tsx
