+++ src/i18n/messages/ar.json
@@ -166,5 +166,6 @@
   "review": "المراجعة",
   "enter-the-title": "أدخل العنوان",
   "what-do-you-think-of-this-product": "ما رأيك في هذا المنتج؟",
-  "add-review": "إضافة المراجعة"
+  "add-review": "إضافة المراجعة",
+  "login-to-add-review": "يرجى تسجيل الدخول لتتمكن من إضافة تقييم إلى المنتج\t"
 }
